Install codetools via r-recommended.
codetools is needed to use knitr's dependson caching feature in the tests. codetools is only suggested by knitr, but it is included in most R installations (b/c it is a recommended package). Thus it makes sense to install it here than to add it to the Suggests of workflowr (which never actually uses it).

Also add qpdf to remove that warning.