Remove littler. Doesn't respect R_LIBS setting.
None of the packages were getting cached because it ignored this.

Strangely it worked for some of the other builds, so not sure
exactly what is going on. My guess is that this is due to the
version difference of littler in this image with R 3.2.5.

Also created that spurious error with wflow_build() where the files
were not created. In general I'm tired of troubleshooting errors
caused entirely by littler not behaving like R/Rscript.

Fortunately the usethis CircleCI template I created uses Rscript.