Try manually installing caTools to fix r325 failure.
rmarkdown 1.7 requires caTools, but not 1.8. It's unclear to me  why
caTools can't be found when using install_version(). It should just
install it.

How a user ends up with this setup isn't really the point of the test,
but to make sure workflowr works properly in such a context. Thus I
just need to get everything installed.