Install latest reticulate (1.15, required min version for workflowr).
Also harmonized approach to installing packages in jobs that use
MRAN snapshots. Only packages that weren't available are installed
prior to the normal installation step. Packages that simply need a
later version are installed after this, with no dependencies updated.
This will only cause an issue if a new dependency was added in the new
version that wasn't a dependency of another package, a risk I am willing
to take.