Invalidate package caches again.
Same stringi error as before. I had previously fixed it in the last
commit:

09aa4a93cd8402281f6b3f55b31fcb76331a95b1

But that didn't last long. rmd_1.7, rmd_1.10, pandoc_one, and speed
all failed to install knitr due to the stringi error. Note that the
first 3 jobs all share the same cache.

I'm invalidating all the caches to make it easy to mass update in the
future.