Always install stringi for r-base:latest jobs that share cache.
This is a tough call. stringi takes a long time to compile, and is
one of the main motivations for using a package cache. On the other
hand, the jobs consistently failing is requiring manually updating the
caches anyways, so I'm not saving any machine time, and am definintely
losing human time. Another option would be to use separate caches. Part
of the problem is that the rmd jobs downgrade knitr and rmarkdown. Thus
depending on which cache gets saved first (CircleCI only saves it once),
then each time knitr and rmarkdown need to be re-installed (instead of
only causing a problem when there is a CRAN update). But maintaining
separate caches for these similar setups would take up more space. For
now I'm going with installing stringi since it's straightforward, and
these jobs only run weekly.