Try to debug disappearing htmltools package.
I can clearly see it is installed in the log. But then when rmarkdown
is installed later with no dependencies (to not change anything else),
it can't be found. This error is repeatable. And when I logged in via
SSH, I confirmed that htmltools had disappeared, and that could easily
be installed with install.packages().

Usually this sort of thing happens when the rmarkdown package adds or
removes a dependency, e.g. caTools was dropped in 1.8.

https://github.com/workflowr/workflowr-pkg-tests/commit/32fc2575b948fabfa0ee7c353c1bcfab8a96aa3d

But htmltools is a dependency for 1.7 and the latest (2.5). The only
thing that changed was the minimum, which is met by the MRAN version.

Sometimes these errors are mysterious. I'd like to figure it out so
I can prevent in the future. But worst case I'll simply install it a
second time to ensure it is there.

https://github.com/workflowr/workflowr-pkg-tests/commit/835e71381f09851e962a125279c9e898829dcc9f

The goal is to test workflowr with R 3.2.5, not to debug a crazy situation
with a mix of R packages from an MRAN snapshot and CRAN.