Update minimum version tested from R 3.2.5 to R 3.3.0.
testthat 3.0.1 now requires rlang 0.4.9, which requires R 3.3.0. In
general the tidyverse is dropping support for R 3.2.5. While it might
be possible for me to come up with some strategy for continued testing
with R 3.2.5, it doesn't seem worth the effort. R 3.2.5 and 3.3.0 were
both released in April 2016, and Ubuntu 18.04 LTS installs R 3.4.4, so
I supsect very few people are still using R 3.2.5.

https://github.com/r-lib/testthat/commit/ef07c0d22f3b1e8845f0559ff36df96445ace094
https://github.com/r-lib/rlang/blob/0852ea2db34132d788def70181547cd04ba2bb0d/DESCRIPTION#L17
https://cran.r-project.org/bin/windows/base/old/

Note that this may not work without deleting the cached packages, but
I wanted to try it first.