Explicitly uninstall pandoc for pandocless tests
r-cran-rmarkdown now depends on pandoc. In general I think this is a
good practice for a package manager (we also add pandoc to the run
requirements of the r-rmarkdown conda-forge package for the same
reason). But in order to confirm `R CMD check` can run without pandoc, I
need to remove it.

Can't use install.packages() because that is configured to use APT
via {bpsm}